adding some bits of refactoring of launching server in background factored out
